BUG: Fix VTK file reading and point positions for Slicer
Point positions were previously read line-by-line without respect for the coordinate system that the file was saved in (LPS/RAS).
This commit uses vtkMRMLModelStorageNode to load the points (in RAS), and converts them into LPS.

This removes the need for the IgnoreFirstValue, IntendedUse arguments, as the points should be handled correctly regardless of their intended use.